third_party/WebKit/LayoutTests/sensor/idl-Accelerometer.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  // Test that Accelerometer interface exists
  assert_true('Accelerometer' in window);
}, 'Test that the Accelerometer interface is present');
</script>
